warning fix

+++ cpp_rutils/literals.h
@@ -7,13 +7,13 @@
 inline
 constexpr size_t operator "" _uz(unsigned long long n)
 {
-    return n;
+	return n;
 }
 
 
 inline constexpr unsigned char operator "" _uchar(char arg) noexcept
 {
-    return static_cast<unsigned char>(arg);
+	return static_cast<unsigned char>(arg);
 }
 
 // byte/megabyte/.... literals
@@ -43,9 +43,9 @@ constexpr my_uint64_t operator"" _GB(my_uint64_t giga_bytes)
 }
 inline double operator"" _pct(long double d)
 {
-	return 0.01 * d;
+	return 0.01 * static_cast<double>(d);
 }
 inline double operator"" _bps(long double d)
 {
-	return 0.0001 * d;
+	return 0.0001 * static_cast<double>(d);
 }
